Transform - validate tasks
File with the first portion of the transform task copied onto it. Small spellcheck changes made to the data cleaning portion of the transformation task. Should now fit both halves of the transformation task together, but might still want to rearrange lines, since I mainly copied the data cleaning portion to it. Includes transformation and validation tasks with prototype xcom pushes and pulls. Transformation and validation can be separated if so desired.