New version of transformation and validation tasks
New version of the transformation and validation tasks with the formatted date column added back and the index removed. Changed the whole validation task to only check for missing values, ranges, and outliers, and to print error messages in case of any of them. The validation task will fail if anything in it fails. Will still need to check the critical columns list that all columns actually exist.